refac: move modal to componentd
